Modify Azure, EC2, and GCE plugins to use FQDN modules
Modified Azure, EC2, and GCE plugins to use fully-qualified names for all
Ansible modules. This has been the recommended practice for several years,
and it allows users to run the various 'lint' programs without the requirement
to explicitly disable some of the tests.

(Some of the drivers already include snippets that will disable some of
these tests in the generated code - but that may need to be disabled due
to compliance requirements.)

I also included a small change to 'test_ec2.py' due to 'critical' error
related to an unqualified method name. I didn't make the same change with
GCE - I can use EC2 locally but not GCE.

Local tox results

The local tox results were disappointing - but it is probably due to
omissions in my local environment. The two most noteworthy errors - and
definitely not related to my environment - are:

 - Vagrant failed since the 'python-vagrant' package is not installed.
   That's a python 2.7 module - it should be using python3-vagrant.
   (I verified that I have both vagrant and python3-vagrant installed.)

 - Docker failed due to failure to find the 'containers.podman.podman_network'
   module. I've verified that I have the most recent collection installed
   (1.10.1) and that the method is present.